pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "corridoriq-backend"
version = "0.1.0"
description = "CorridorIQ backend for corridor scoring, diagnosis, and recommendations."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.30,<1.0",
  "pydantic>=2.10,<3.0",
]

[project.optional-dependencies]
pipeline = [
  "pandas>=2.0",
  "python-dotenv>=1.0",
  "requests>=2.31",
  "supabase>=2.0,<3.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "pipeline*"]

